refactor : repository 폴더내로 경로 변경
