+++ src/main/java/com/coffee_shop/coffeeshop/domain/coupon/repository/CouponRepository.java
@@ -1,6 +1,8 @@
-package com.coffee_shop.coffeeshop.domain.coupon;
+package com.coffee_shop.coffeeshop.domain.coupon.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.coffee_shop.coffeeshop.domain.coupon.Coupon;
 
 public interface CouponRepository extends JpaRepository<Coupon, Long> {
 }
