+++ src/main/java/com/coffee_shop/coffeeshop/service/coupon/CouponService.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.coffee_shop.coffeeshop.domain.coupon.Coupon;
-import com.coffee_shop.coffeeshop.domain.coupon.CouponRepository;
+import com.coffee_shop.coffeeshop.domain.coupon.repository.CouponRepository;
 import com.coffee_shop.coffeeshop.service.coupon.dto.request.CouponSaveServiceRequest;
 
 import lombok.RequiredArgsConstructor;
